+++ DesktopApplicationTemplate.UI/ViewModels/AdvancedConfigViewModelBase.cs
@@ -1,7 +1,7 @@
 using System;
 using System.Windows.Input;
 using DesktopApplicationTemplate.UI.Helpers;
-using DesktopApplicationTemplate.UI.Services;
+using DesktopApplicationTemplate.Core.Services;
 
 namespace DesktopApplicationTemplate.UI.ViewModels;