+++ DesktopApplicationTemplate.UI/ViewModels/ServiceCreateViewModelBase.cs
@@ -1,5 +1,4 @@
 using System.Windows.Input;
-using DesktopApplicationTemplate.UI.Helpers;
 using DesktopApplicationTemplate.Core.Services;
 
 namespace DesktopApplicationTemplate.UI.ViewModels;
@@ -8,58 +7,19 @@ namespace DesktopApplicationTemplate.UI.ViewModels;
 /// Provides shared command wiring for service creation view models.
 /// </summary>
 /// <typeparam name="TOptions">Type of options managed by the service.</typeparam>
-public abstract class ServiceCreateViewModelBase<TOptions> : ValidatableViewModelBase, ILoggingViewModel
+public abstract class ServiceCreateViewModelBase<TOptions> : ServiceEditorViewModelBase<TOptions>
 {
     /// <summary>
     /// Initializes a new instance of the <see cref="ServiceCreateViewModelBase{TOptions}"/> class.
     /// </summary>
-    /// <param name="logger">Optional logging service.</param>
-    protected ServiceCreateViewModelBase(ILoggingService? logger = null)
+    protected ServiceCreateViewModelBase(IServiceRule rule, ILoggingService? logger = null)
+        : base(rule, logger)
     {
-        Logger = logger;
-        var saveCommand = new RelayCommand(OnSave, () => !HasErrors);
-        SaveCommand = saveCommand;
-        CancelCommand = new RelayCommand(OnCancel);
-        AdvancedConfigCommand = new RelayCommand(OnAdvancedConfig);
-        ErrorsChanged += (_, _) => saveCommand.RaiseCanExecuteChanged();
+        SaveButtonText = "Create";
     }
 
-    /// <inheritdoc />
-    public ILoggingService? Logger { get; set; }
-
     /// <summary>
-    /// Command executed to save the service.
-    /// </summary>
-    public ICommand SaveCommand { get; }
-
-    /// <summary>
-    /// Alias for <see cref="SaveCommand"/> used by XAML bindings.
+    /// Alias for <see cref="ServiceEditorViewModelBase{TOptions}.SaveCommand"/> used by XAML bindings.
     /// </summary>
     public ICommand CreateCommand => SaveCommand;
-
-    /// <summary>
-    /// Command executed to cancel creation.
-    /// </summary>
-    public ICommand CancelCommand { get; }
-
-    /// <summary>
-    /// Command executed to open advanced configuration.
-    /// </summary>
-    public ICommand AdvancedConfigCommand { get; }
-
-    /// <summary>
-    /// Handles save operations for the service.
-    /// </summary>
-    protected abstract void OnSave();
-
-    /// <summary>
-    /// Handles cancellation of the operation.
-    /// </summary>
-    protected abstract void OnCancel();
-
-    /// <summary>
-    /// Handles requests for advanced configuration.
-    /// </summary>
-    protected abstract void OnAdvancedConfig();
 }
-