+++ DesktopApplicationTemplate.UI/ViewModels/TcpCreateServiceViewModel.cs
@@ -9,8 +9,6 @@ namespace DesktopApplicationTemplate.UI.ViewModels;
 /// </summary>
 public class TcpCreateServiceViewModel : ServiceCreateViewModelBase<TcpServiceOptions>
 {
-    private readonly IServiceRule _rule;
-    private string _serviceName = string.Empty;
     private string _host = string.Empty;
     private int _port;
 
@@ -23,38 +21,8 @@ public class TcpCreateServiceViewModel : ServiceCreateViewModelBase<TcpServiceOp
     /// Initializes a new instance of the <see cref="TcpCreateServiceViewModel"/> class.
     /// </summary>
     public TcpCreateServiceViewModel(IServiceRule rule, ILoggingService? logger = null)
-        : base(logger)
+        : base(rule, logger)
     {
-        _rule = rule ?? throw new ArgumentNullException(nameof(rule));
-    }
-
-    /// <summary>
-    /// Raised when the user finishes configuring the service.
-    /// </summary>
-    public event Action<string, TcpServiceOptions>? ServiceCreated;
-
-    /// <summary>
-    /// Raised when the user cancels configuration.
-    /// </summary>
-    public event Action? Cancelled;
-
-
-    /// <summary>
-    /// Name of the service to create.
-    /// </summary>
-    public string ServiceName
-    {
-        get => _serviceName;
-        set
-        {
-            _serviceName = value;
-            var error = _rule.ValidateRequired(value, "Service name");
-            if (error is not null)
-                AddError(nameof(ServiceName), error);
-            else
-                ClearErrors(nameof(ServiceName));
-            OnPropertyChanged();
-        }
     }
 
     /// <summary>
@@ -66,7 +34,7 @@ public class TcpCreateServiceViewModel : ServiceCreateViewModelBase<TcpServiceOp
         set
         {
             _host = value;
-            var error = _rule.ValidateRequired(value, "Host");
+            var error = Rule.ValidateRequired(value, "Host");
             if (error is not null)
                 AddError(nameof(Host), error);
             else
@@ -84,7 +52,7 @@ public class TcpCreateServiceViewModel : ServiceCreateViewModelBase<TcpServiceOp
         set
         {
             _port = value;
-            var error = _rule.ValidatePort(value);
+            var error = Rule.ValidatePort(value);
             if (error is not null)
                 AddError(nameof(Port), error);
             else
@@ -92,11 +60,6 @@ public class TcpCreateServiceViewModel : ServiceCreateViewModelBase<TcpServiceOp
             OnPropertyChanged();
         }
     }
-
-    /// <summary>
-    /// Raised when advanced configuration is requested.
-    /// </summary>
-    public event Action<TcpServiceOptions>? AdvancedConfigRequested;
 
     /// <inheritdoc />
     protected override void OnSave()
@@ -110,14 +73,14 @@ public class TcpCreateServiceViewModel : ServiceCreateViewModelBase<TcpServiceOp
         Options.Host = Host;
         Options.Port = Port;
         Logger?.Log("TCP create options finished", LogLevel.Debug);
-        ServiceCreated?.Invoke(ServiceName, Options);
+        RaiseServiceSaved(Options);
     }
 
     /// <inheritdoc />
     protected override void OnCancel()
     {
         Logger?.Log("TCP create options cancelled", LogLevel.Debug);
-        Cancelled?.Invoke();
+        RaiseEditCancelled();
     }
 
     /// <inheritdoc />
@@ -126,6 +89,6 @@ public class TcpCreateServiceViewModel : ServiceCreateViewModelBase<TcpServiceOp
         Logger?.Log("Opening TCP advanced config", LogLevel.Debug);
         Options.Host = Host;
         Options.Port = Port;
-        AdvancedConfigRequested?.Invoke(Options);
+        RaiseAdvancedConfigRequested(Options);
     }
 }